♻️ Refactor app entry point code
