+++ Dockerfile
@@ -15,7 +15,6 @@ WORKDIR /home/operations-engineering-join-github
 
 COPY requirements.txt requirements.txt
 COPY app app
-COPY operations_engineering_join_github.py operations_engineering_join_github.py
 
 RUN pip3 install --no-cache-dir -r requirements.txt
 
@@ -28,4 +27,4 @@ EXPOSE 4567
 
 HEALTHCHECK --interval=60s --timeout=30s CMD curl -I -XGET http://localhost:4567 || exit 1
 
-ENTRYPOINT ["gunicorn", "--bind=0.0.0.0:4567", "operations_engineering_join_github:app()"]
+ENTRYPOINT ["gunicorn", "--bind=0.0.0.0:4567", "app.run:app()"]
