--- setup.cfg
@@ -1,10 +1,0 @@
-[flake8]
-max-line-length = 150
-ignore = E501,W503
-exclude =  .git, .eggs, __pycache__, env
-
-[mypy]
-ignore_missing_imports = true
-
-[pylint.FORMAT]
-max-line-length = 150
